WIP: Nuklear

+++ Engine/src/teleios/memory/pool.h
@@ -7,6 +7,6 @@
 TLMemoryPool* tl_memory_pool_create(void);
 void tl_memory_pool_destroy(TLMemoryPool* pool);
 
-void* tl_memory_pool_alloc(TLMemoryPool* pool, u64 size);
+void* tl_memory_pool_alloc(TLMemoryPool* pool, u32 size);
 
 #endif // TELEIOS_MEMORY_POOL_H
